+++ lib/app/app.dart
@@ -1,3 +1,2 @@
-export 'config.dart';
 export 'di.dart';
 export 'guess_song_app.dart';