+++ lib/app/di.dart
@@ -2,7 +2,7 @@ import 'package:get_it/get_it.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
-import '../api/dio_client.dart';
+import '../api/itunes_dio_client.dart';
 
 final G = GetIt.instance;
 
@@ -12,9 +12,10 @@ abstract class DIContainer {
     talker.debug('Initializing the application...');
 
     final sharedPreferences = await SharedPreferences.getInstance();
-    // final apiClient = YoutubeDioClient(talker);
+    final iTunesApiClient = ITunesApiDioClient(talker);
 
     G.registerSingleton<Talker>(talker);
     G.registerSingleton<SharedPreferences>(sharedPreferences);
+    G.registerSingleton<ITunesApiDioClient>(iTunesApiClient);
   }
 }