+++ lib/features/game_flow/models/question.dart
@@ -1,0 +1,8 @@
+import '../../../repositories/music_api/models/song.dart';
+
+class Question {
+  Question({required this.song, required this.options});
+
+  final Song song;
+  final List<String> options;
+}