+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 import 'package:talker_riverpod_logger/talker_riverpod_logger.dart';
@@ -9,7 +8,6 @@ import 'package:talker_riverpod_logger/talker_riverpod_logger.dart';
 import 'app/app.dart';
 
 Future<void> main() async {
-  // await dotenv.load(fileName: ".env");
   await DIContainer.init();
 
   runZonedGuarded(() {