+++ lib/repositories/local_storage/local_storage_interface.dart
@@ -1,0 +1,4 @@
+abstract interface class ILocalStorage {
+  String? getUsername();
+  Future<void> saveUsername(String username);
+}