+++ lib/repositories/music_api/music_api.dart
@@ -1,0 +1,2 @@
+export 'itunes_repository.dart';
+export 'music_api_interface.dart';