+++ lib/repositories/music_api/music_api_interface.dart
@@ -1,0 +1,6 @@
+import 'models/song.dart';
+
+abstract interface class IMusicRepository {
+  Future<List<Song>> getTracksByArtist(String artistName);
+  Future<List<Song>> getTracksFromAlbum({required String artistName, required String albumName});
+}