+++ lib/repositories/theme/theme.dart
@@ -1,2 +1,2 @@
+export 'theme_interface.dart';
 export 'theme_repository.dart';
-export 'theme_repository_interface.dart';